Create own file for NavigationLinkAppender and add to submenu navigation component
